do not overwrite unchanged objects when restoring bucket
When a bucket is restored to an earlier state of itself, only a few objects
might actually have changed. Changed objects can be identified by comparing
the etag of the desired version with the etag of the most recent version.
If the etags agree, the object can be skipped. Because the most recent version
of any object is always listed first in the response from list_object_versions,
its etag is definitely known when handling the version to be restored.

Added a test that checks that no new version is created when the object was
not changed.

Also running the existing test in a configuration that uses the same source
and destination bucket without requiring another call of
"s3-pit-restore --test". Because this is a frequent use case, it should be
tested without special effort.